Moved API methods into a more organized structure
