+++ communications/tmi.py
@@ -1,0 +1,6 @@
+import json
+import requests
+from utils import cnf
+
+class TMI:
+    pass
